requirements.txt
boto==2.38.0
python-dateutil==2.4.2